Improved Face Recognition by adding better parameters for face detection Renamed files name with uppercase characters to full lowercase name Added change VideoArea image function on pressing 'E' key Added labels with information about faces database in other csv file(labels.csv) Started adding graphic information about recognized faces

+++ csv_handler.h
@@ -42,7 +42,7 @@ namespace CsvHandler {
         return true;
     }
     
-    bool writeCsvFile(const string &filename, const char &separator, const int &nCols, const vector<string*> &lines){
+    bool writeCsvFile(const string &filename, const char &separator, const size_t &nCols, const vector<string*> &lines){
         ofstream file( filename, ofstream::out );
         if (!file.good()){
             return false;
